Added a script which turns new lines into '\n'
New lines will turn into a backslash followed by n. Useful for copying text into programs for a string. Will take STDIN or a file as input then output as STDOUT or into the provided file name. If only one file name is given, it will test if it is file, if it isn't, it will read from STDIN. (This could cause errors if piping. If piping, send it to STDOUT and redirect if you need it in a file.)
Outputs from STDOUT will end with a newline for readability.